feature(grids) Size Label Tweaks - Added `controlForStrategy` so memoized controls can be used. - Added `StrategySizeLabel` to every strategy that included `AbsoluteResizeControl`. - Include `SizeLabel` in `NewCanvasControlsInner` if `StrategySizeLabel`   hasn't been included in the strategy controls. - Extracted out `SizeLabel` into its own file. - `sizeLabel` function now supports `CSSNumber` for the `state` parameter. - `sizeLabelContents` now tries to present the unit values for the dimensions.
