+++ editor/src/components/canvas/controls/select-mode/absolute-resize-control.tsx
@@ -1,28 +1,18 @@
 import React from 'react'
 import { MetadataUtils } from '../../../../core/model/element-metadata-utils'
 import * as EP from '../../../../core/shared/element-path'
-import type { ElementInstanceMetadataMap } from '../../../../core/shared/element-template'
-import type { CanvasRectangle, CanvasVector } from '../../../../core/shared/math-utils'
-import {
-  boundingRectangleArray,
-  canvasRectangle,
-  isInfinityRectangle,
-  nullIfInfinity,
-  windowPoint,
-} from '../../../../core/shared/math-utils'
+import type { CanvasVector } from '../../../../core/shared/math-utils'
+import { windowPoint } from '../../../../core/shared/math-utils'
 import type { ElementPath } from '../../../../core/shared/project-file-types'
-import { assertNever, NO_OP } from '../../../../core/shared/utils'
+import { NO_OP } from '../../../../core/shared/utils'
 import { Modifier } from '../../../../utils/modifiers'
 import { when } from '../../../../utils/react-conditionals'
 import { useColorTheme } from '../../../../uuiui'
 import type { EditorDispatch } from '../../../editor/action-types'
 import { applyCommandsAction } from '../../../editor/actions/action-creators'
 import { useDispatch } from '../../../editor/store/dispatch-context'
-import { getMetadata } from '../../../editor/store/editor-state'
 import { Substores, useEditorState, useRefEditorState } from '../../../editor/store/store-hook'
-import type { FixedHugFill } from '../../../inspector/inspector-common'
 import {
-  detectFillHugFixedState,
   invert,
   isFillOrStretchModeAppliedOnAnySide,
   isFillOrStretchModeAppliedOnSpecificSide,
@@ -35,13 +25,12 @@ import { controlForStrategyMemoized } from '../../canvas-strategies/canvas-strat
 import { createInteractionViaMouse } from '../../canvas-strategies/interaction-state'
 import type { EdgePosition } from '../../canvas-types'
 import { CSSCursor } from '../../canvas-types'
-import { getAllTargetsUnderAreaAABB, windowToCanvasCoordinates } from '../../dom-lookup'
+import { windowToCanvasCoordinates } from '../../dom-lookup'
 import { useBoundingBox } from '../bounding-box-hooks'
 import { CanvasOffsetWrapper } from '../canvas-offset-wrapper'
 import { isZeroSizedElement } from '../outline-utils'
 import { useMaybeHighlightElement } from './select-mode-hooks'
 import { isEdgePositionEqualTo } from '../../canvas-utils'
-import { treatElementAsGroupLike } from '../../canvas-strategies/strategies/group-helpers'
 import { useResizeEdges } from './use-resize-edges'
 
 export const AbsoluteResizeControlTestId = (targets: Array<ElementPath>): string =>
@@ -51,9 +40,6 @@ interface AbsoluteResizeControlProps {
   targets: Array<ElementPath>
   pathsWereReplaced: boolean
 }
-
-export const SizeLabelID = 'SizeLabel'
-export const SizeLabelTestId = 'SizeLabelTestId'
 
 export const AbsoluteResizeControl = controlForStrategyMemoized(
   ({ targets, pathsWereReplaced }: AbsoluteResizeControlProps) => {
@@ -88,12 +74,6 @@ export const AbsoluteResizeControl = controlForStrategyMemoized(
     const bottomRightRef = useBoundingBox(targets, (ref, boundingBox) => {
       ref.current.style.left = boundingBox.width + 'px'
       ref.current.style.top = boundingBox.height + 'px'
-    })
-
-    const resizeRef = useBoundingBox(targets, (ref, boundingBox) => {
-      ref.current.style.top = boundingBox.height + 'px'
-      ref.current.style.left = 0 + 'px'
-      ref.current.style.width = boundingBox.width + 'px'
     })
 
     const scale = useEditorState(
@@ -218,7 +198,6 @@ export const AbsoluteResizeControl = controlForStrategyMemoized(
               />
             </React.Fragment>,
           )}
-          <SizeLabel ref={resizeRef} targets={targets} pathsWereReplaced={pathsWereReplaced} />
         </div>
       </CanvasOffsetWrapper>
     )
@@ -342,252 +321,6 @@ const ResizePoint = React.memo(
   }),
 )
 ResizePoint.displayName = 'ResizePoint'
-
-const sizeLabel = (state: FixedHugFill['type'], actualSize: number): string => {
-  switch (state) {
-    case 'fill':
-      return 'Fill'
-    case 'hug':
-    case 'hug-group':
-      return 'Hug'
-    case 'squeeze':
-      return 'Squeeze'
-    case 'collapsed':
-      return 'Collapsed'
-    case 'fixed':
-    case 'scaled':
-    case 'detected':
-    case 'computed':
-      return `${actualSize}`
-    case 'stretch':
-      return 'Stretch'
-    default:
-      assertNever(state)
-  }
-}
-
-export type SizeLabelSize = {
-  type: 'SIZE_LABEL_WITH_DIMENSIONS'
-  h: string
-  v: string
-}
-
-function sizeLabelWithDimensions(h: string, v: string): SizeLabelSize {
-  return {
-    type: 'SIZE_LABEL_WITH_DIMENSIONS',
-    h: h,
-    v: v,
-  }
-}
-
-export type SizeLabelGroup = {
-  type: 'SIZE_LABEL_GROUP'
-}
-
-function sizeLabelGroup(): SizeLabelGroup {
-  return {
-    type: 'SIZE_LABEL_GROUP',
-  }
-}
-
-export type SizeLabelChildren = {
-  type: 'SIZE_LABEL_CHILDREN'
-  h: string
-  v: string
-}
-
-function sizeLabelChildren(h: string, v: string): SizeLabelChildren {
-  return {
-    type: 'SIZE_LABEL_CHILDREN',
-    h: h,
-    v: v,
-  }
-}
-
-export type SizeLabelContents = SizeLabelSize | SizeLabelGroup | SizeLabelChildren
-
-function sizeLabelContents(
-  metadata: ElementInstanceMetadataMap,
-  selectedElements: Array<ElementPath>,
-  boundingBox: CanvasRectangle | null,
-  pathsWereReplaced: boolean,
-): SizeLabelContents | null {
-  if (selectedElements.length === 0) {
-    return null
-  }
-
-  if (selectedElements.length === 1) {
-    if (treatElementAsGroupLike(metadata, selectedElements[0])) {
-      return sizeLabelGroup()
-    }
-
-    const globalFrame = MetadataUtils.findElementByElementPath(
-      metadata,
-      selectedElements[0],
-    )?.globalFrame
-    if (globalFrame == null || isInfinityRectangle(globalFrame)) {
-      return null
-    }
-
-    const horizontal =
-      detectFillHugFixedState('horizontal', metadata, selectedElements[0]).fixedHugFill?.type ??
-      'fixed'
-    const vertical =
-      detectFillHugFixedState('vertical', metadata, selectedElements[0]).fixedHugFill?.type ??
-      'fixed'
-
-    if (pathsWereReplaced) {
-      return sizeLabelChildren(
-        sizeLabel(horizontal, globalFrame.width),
-        sizeLabel(vertical, globalFrame.height),
-      )
-    }
-
-    return sizeLabelWithDimensions(
-      sizeLabel(horizontal, globalFrame.width),
-      sizeLabel(vertical, globalFrame.height),
-    )
-  }
-
-  if (boundingBox != null) {
-    return sizeLabelWithDimensions(`${boundingBox.width}`, `${boundingBox.height}`)
-  }
-
-  return null
-}
-
-interface SizeLabelProps {
-  targets: Array<ElementPath>
-  pathsWereReplaced: boolean
-}
-
-const FontSize = 11
-const PaddingV = 0
-const PaddingH = 2
-const ExplicitHeightHacked = 20
-const BorderRadius = 2
-const SizeLabelMarginTop = 8
-
-function getLabelText(label: SizeLabelContents | null): string | null {
-  if (label == null) {
-    return null
-  }
-  switch (label.type) {
-    case 'SIZE_LABEL_GROUP':
-      return 'Group'
-    case 'SIZE_LABEL_WITH_DIMENSIONS':
-      return `${label.h} x ${label.v}`
-    case 'SIZE_LABEL_CHILDREN':
-      return `(Children) ${label.h} x ${label.v}`
-    default:
-      assertNever(label)
-  }
-}
-
-const SizeLabel = React.memo(
-  React.forwardRef<HTMLDivElement, SizeLabelProps>(({ targets, pathsWereReplaced }, ref) => {
-    const scale = useEditorState(
-      Substores.canvas,
-      (store) => store.editor.canvas.scale,
-      'Resizelabel scale',
-    )
-    const colorTheme = useColorTheme()
-    const metadata = useEditorState(
-      Substores.metadata,
-      (store) => getMetadata(store.editor),
-      'ResizeLabel metadata',
-    )
-
-    const boundingBox = boundingRectangleArray(
-      targets.map((t) => nullIfInfinity(MetadataUtils.getFrameInCanvasCoords(t, metadata))),
-    )
-
-    const label = sizeLabelContents(metadata, targets, boundingBox, pathsWereReplaced)
-    const labelText = getLabelText(label)
-
-    const [dimmed, setDimmed] = React.useState(false)
-
-    const editorRef = useRefEditorState((store) => ({
-      scale: store.editor.canvas.scale,
-      offset: store.editor.canvas.roundedCanvasOffset,
-      jsxMetadata: store.editor.jsxMetadata,
-      elementPathTree: store.editor.elementPathTree,
-      allElementProps: store.editor.allElementProps,
-      hiddenInstances: store.editor.hiddenInstances,
-    }))
-
-    const onMouseEnter = React.useCallback(() => {
-      const distanceBetweenBoxAndLabel = 10 // px
-      const labelRect = document.getElementById(SizeLabelID)?.getBoundingClientRect()
-      if (boundingBox != null && labelRect != null) {
-        const area = canvasRectangle({
-          x: boundingBox.x + (boundingBox.width - labelRect.width) / 2,
-          y:
-            boundingBox.y +
-            boundingBox.height +
-            distanceBetweenBoxAndLabel / editorRef.current.scale,
-          width: labelRect.width,
-          height: labelRect.height,
-        })
-        const elementsUnderLabel = getAllTargetsUnderAreaAABB(
-          editorRef.current.jsxMetadata,
-          [],
-          editorRef.current.hiddenInstances,
-          'no-filter',
-          area,
-          editorRef.current.elementPathTree,
-          editorRef.current.allElementProps,
-          true,
-        )
-        setDimmed(elementsUnderLabel.length > 0)
-      }
-    }, [editorRef, boundingBox])
-
-    const onMouseLeave = React.useCallback(() => {
-      setDimmed(false)
-    }, [])
-
-    return (
-      <div
-        ref={ref}
-        style={{
-          position: 'absolute',
-          display: 'flex',
-          justifyContent: 'center',
-          pointerEvents: 'none',
-        }}
-        data-testid='parent-resize-label'
-      >
-        {when(
-          labelText != null,
-          <div
-            id={SizeLabelID}
-            data-testid={SizeLabelTestId}
-            style={{
-              display: 'flex',
-              alignItems: 'center',
-              marginTop: SizeLabelMarginTop / scale,
-              padding: `${PaddingV}px ${PaddingH / scale}px`,
-              borderRadius: BorderRadius / scale,
-              color: colorTheme.white.value,
-              backgroundColor: colorTheme.primary.value,
-              fontSize: FontSize / scale,
-              height: ExplicitHeightHacked / scale,
-              opacity: dimmed ? 0.075 : 1,
-              transition: '0.1s',
-              pointerEvents: 'initial',
-            }}
-            onMouseEnter={onMouseEnter}
-            onMouseLeave={onMouseLeave}
-          >
-            {labelText}
-          </div>,
-        )}
-      </div>
-    )
-  }),
-)
-SizeLabel.displayName = 'SizeLabel'
 
 function startResizeInteraction(
   event: React.MouseEvent<HTMLDivElement>,
